requirements.txt
numpy
opencv-contrib-python
scipy
matplotlib
vtk
pupil-apriltags
gtsam
filterpy
pyntcore
scikit-learn
